refactor: moved compile_regex in utils
